+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.76.6
 
+- The backup validator accepts the new "forward secrecy" format as well as the existing "ciphertext only" format.